+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
 ### New in 0.46 (not released yet)
 
-* _Nothing yet_
+* Fix: EventFlow now uses a Autofac lifetime scope for validating service
+  registrations when `IEventFlowOpions.CreateResolver(true)` is invoked.
+  Previously services were created but never disposed as they were resolved
+  using the root container
 
 ### New in 0.45.2877 (released 2017-05-28)